+++ client/store/index.js
@@ -10,8 +10,9 @@ import carts from './cart'
 import reviews from './reviews'
 import users from './users'
 import cart from './cart'
+import cartItem from './cartItem'
 
-const reducer = combineReducers({ user, products, categories, reviews, users, cart })
+const reducer = combineReducers({ user, products, categories, reviews, users, cart, cartItem })
 const middleware = composeWithDevTools(applyMiddleware(
   thunkMiddleware,
   createLogger({ collapsed: true })
@@ -25,3 +26,4 @@ export * from './category'
 export * from './users'
 export * from './reviews'
 export * from './cart'
+export * from './cartItem'
